fix: move tooltip setup to App
